hssi: move pkt filt programming to hssi_cmd.h
Implement the pkt filt dest address setter as a member function
of the hssi_cmd class so that it can be shared between each of
the test modes.
