+++ samples/hssi/hssi_cmd.h
@@ -1,4 +1,4 @@
-// Copyright(c) 2020, Intel Corporation
+// Copyright(c) 2020-2021, Intel Corporation
 //
 // Redistribution  and  use  in source  and  binary  forms,  with  or  without
 // modification, are permitted provided that the following conditions are met:
@@ -30,8 +30,11 @@
 #include <cstring>
 #include <netinet/ether.h>
 #include "afu_test.h"
+#include <opae/uio.h>
 
 using test_command = opae::afu_test::command;
+
+#define PKT_FILT_CSR_DEST_ADDR 0x0040
 
 #define INVALID_MAC 0xffffffffffffffffULL
 
@@ -41,6 +44,39 @@ public:
   hssi_cmd()
   : running_(true)
   {}
+
+  int set_pkt_filt_dest(const std::string &dfl_dev,
+                        uint64_t bin_dest_addr) const
+  {
+    struct opae_uio uio;
+
+    if (opae_uio_open(&uio, dfl_dev.c_str())) {
+      std::cerr << "Failed to open: " << dfl_dev << std::endl;
+      return test_afu::error;
+    }
+
+    uint8_t *mmio = nullptr;
+
+    if (opae_uio_region_get(&uio, 0, &mmio, nullptr)) {
+      std::cerr << "Failed to get region 0 from: " << dfl_dev << std::endl;
+      opae_uio_close(&uio);
+      return test_afu::error;
+    }
+
+    *(volatile uint64_t *)(mmio + PKT_FILT_CSR_DEST_ADDR) = bin_dest_addr;
+    uint64_t dest_addr = *(volatile uint64_t *)(mmio + PKT_FILT_CSR_DEST_ADDR);
+
+    if (bin_dest_addr != dest_addr) {
+      std::cerr << "Packet Filter dest address mismatch! wrote: " << int_to_hex(bin_dest_addr)
+                << " read: " << int_to_hex(dest_addr) << std::endl;
+      opae_uio_close(&uio);
+      return test_afu::error;
+    }
+
+    opae_uio_close(&uio);
+
+    return test_afu::success;
+  }
 
   uint64_t mac_bits_for(std::string addr) const
   {
